backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# AI and ML
openai==1.3.5
anthropic==0.39.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.2
scikit-image==0.22.0
imutils==0.5.4

# Data processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
aiofiles==23.2.1

# CORS and Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
